+++ src/cold.rs
@@ -1,6 +1,0 @@
-// #[cold]
-// #[inline(always)]
-// #[doc(hidden)]
-// pub const fn err<T, E>(err: E) -> Result<T, E> {
-//     Err(err)
-// }