+++ src/primitive.rs
@@ -7,20 +7,21 @@ use crate::{
     serialize::{write_bytes, Serialize, Sizes},
 };
 
-#[cold]
-#[inline(always)]
-#[doc(hidden)]
-fn cold() {}
-
 macro_rules! impl_primitive {
     () => {};
 
-    ($head:ty $(, $tail:ty)* $(,)?) => {
-        impl_primitive!(@ $head);
-        impl_primitive!($($tail,)*);
+    ([$($head:ident)+] $([$($tail:ident)+])*) => {
+        impl_primitive!(@ < $($head)+);
+        impl_primitive!($([$($tail)+])*);
     };
 
-    (@ $ty:ty) => {
+    (@ $($head:ident)* <) => {};
+    (@ $($head:ident)* < $cursor:ident $($tail:ident)*) => {
+        impl_primitive!{! $($head)* < $cursor < $($tail)* }
+        impl_primitive!{@ $($head)* $cursor < $($tail)* }
+    };
+
+    (! $($from:ident)* < $ty:ident < $($to:ident)*) => {
         impl Formula for $ty {
             const MAX_STACK_SIZE: Option<usize> = Some(size_of::<$ty>());
             const EXACT_SIZE: bool = true;
@@ -44,6 +45,23 @@ macro_rules! impl_primitive {
             }
         }
 
+        $(
+            impl Serialize<$ty> for $from {
+                #[inline(always)]
+                fn serialize<B>(self, sizes: &mut Sizes, buffer: B) -> Result<(), B::Error>
+                where
+                    B: Buffer,
+                {
+                    write_bytes(&$ty::from(self).to_le_bytes(), sizes, buffer)
+                }
+
+                #[inline(always)]
+                fn size_hint(&self) -> Option<Sizes> {
+                    Some(Sizes{ heap: 0, stack: size_of::<$ty>()})
+                }
+            }
+        )*
+
         impl Serialize<$ty> for &$ty {
             #[inline(always)]
             fn serialize<B>(self, sizes: &mut Sizes, buffer: B) -> Result<(), B::Error>
@@ -59,26 +77,35 @@ macro_rules! impl_primitive {
             }
         }
 
+        $(
+            impl Serialize<$ty> for &$from {
+                #[inline(always)]
+                fn serialize<B>(self, sizes: &mut Sizes, buffer: B) -> Result<(), B::Error>
+                where
+                    B: Buffer,
+                {
+                    write_bytes(&$ty::from(*self).to_le_bytes(), sizes, buffer)
+                }
+
+                #[inline(always)]
+                fn size_hint(&self) -> Option<Sizes> {
+                    Some(Sizes{ heap: 0, stack: size_of::<$ty>()})
+                }
+            }
+        )*
+
         impl<T> Deserialize<'_, $ty> for T
         where
             T: From<$ty>,
         {
             #[inline(always)]
-            fn deserialize(de: Deserializer) -> Result<Self, DeserializeError> {
-                let input = de.read_all_bytes();
-                if input.len() == size_of::<$ty>() {
-                    let mut bytes = [0; size_of::<$ty>()];
-                    bytes.copy_from_slice(input);
-                    let value = <$ty>::from_le_bytes(bytes);
-                    return Ok(From::from(value));
-                }
-
-                cold();
-                if input.len() > size_of::<$ty>() {
-                    Err(DeserializeError::WrongLength)
-                } else {
-                    Err(DeserializeError::OutOfBounds)
-                }
+            fn deserialize(mut de: Deserializer) -> Result<Self, DeserializeError> {
+                let input = de.read_bytes(size_of::<$ty>())?;
+                de.finish()?;
+                let mut bytes = [0; size_of::<$ty>()];
+                bytes.copy_from_slice(input);
+                let value = <$ty>::from_le_bytes(bytes);
+                return Ok(From::from(value));
             }
 
             #[inline(always)]
@@ -92,18 +119,9 @@ macro_rules! impl_primitive {
 }
 
 impl_primitive! {
-    u8,
-    u16,
-    u32,
-    u64,
-    u128,
-    i8,
-    i16,
-    i32,
-    i64,
-    i128,
-    f32,
-    f64,
+    [u8 u16 u32 u64 u128]
+    [i8 i16 i32 i64 i128]
+    [f32 f64]
 }
 
 impl Formula for bool {